+++ ReleaseNotes.md
@@ -1,5 +1,10 @@
 # 3.2.3
 
+- Fixed a bug where Buildalyzer.Workspaces would not add transitive references to projects in the Roslyn workspace (#181).
+- No longer attempts to add F# projects from a solution to the Roslyn workspace in Buildalyzer.Workspaces which causes an exception in Roslyn.
+- Updated structure logging support to the latest package providing binlog version 14 parsing (#184).
+- Updated Microsoft.CodeAnalysis package references in Buildalyzer.Workspaces to 3.11.0.
+
 # 3.2.2
 
 - Fixed a bug with WPF custom control libraries (#178, #179, thanks @markrendle)